+++ libhoomd/computes/CellList.cc
@@ -54,6 +54,7 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
 #include "CellList.h"
+#include "Communicator.h"
 
 #include <boost/python.hpp>
 #include <boost/bind.hpp>
@@ -66,7 +67,7 @@ using namespace std;
 /*! \param sysdef system to compute the cell list of
 */
 CellList::CellList(boost::shared_ptr<SystemDefinition> sysdef)
-    : Compute(sysdef),  m_nominal_width(Scalar(1.0)), m_radius(1), m_max_cells(UINT_MAX), m_compute_tdb(false),
+    : Compute(sysdef),  m_nominal_width(Scalar(1.0)), m_radius(1), m_compute_tdb(false),
       m_compute_orientation(false), m_compute_idx(false), m_flag_charge(false), m_flag_type(false), m_sort_cell_list(false)
     {
     m_exec_conf->msg->notice(5) << "Constructing CellList" << endl;
@@ -83,7 +84,7 @@ CellList::CellList(boost::shared_ptr<SystemDefinition> sysdef)
     m_conditions.swap(conditions);
     resetConditions();
 
-    m_num_ghost_cells = make_uint3(0,0,0);
+    m_actual_width = make_scalar3(0.0,0.0,0.0);
     m_ghost_width = make_scalar3(0.0,0.0,0.0);
 
     m_sort_connection = m_pdata->connectParticleSort(bind(&CellList::slotParticlesSorted, this));
@@ -109,7 +110,7 @@ static unsigned int roundDown(unsigned int v, unsigned int m)
     return d*m;
     }
 
-/*! \returns Cell dimensions that match with the current width, box dimension, and max_cells setting
+/*! \returns Cell dimensions that match with the current width, and box dimension
 */
 uint3 CellList::computeDimensions()
     {
@@ -119,44 +120,29 @@ uint3 CellList::computeDimensions()
     const BoxDim& box = m_pdata->getBox();
 
     Scalar3 L = box.getNearestPlaneDistance();
+
     dim.x = roundDown((unsigned int)((L.x) / (m_nominal_width)), m_multiple);
     dim.y = roundDown((unsigned int)((L.y) / (m_nominal_width)), m_multiple);
+    dim.z = (m_sysdef->getNDimensions() == 3) ? roundDown((unsigned int)((L.z) / (m_nominal_width)), m_multiple) : 1;
 
-    // Add a ghost layer on every side where boundary conditions are non-periodic
-    if (! box.getPeriodic().x)
-        dim.x += 2;
-    if (! box.getPeriodic().y)
-        dim.y += 2;
-
-    if (m_sysdef->getNDimensions() == 2)
+    // expand for ghost width if communicating ghosts
+#ifdef ENABLE_MPI
+    if (m_comm)
         {
-        dim.z = 1;
+        const Scalar3 cell_size = make_scalar3(L.x / Scalar(dim.x), L.y / Scalar(dim.y), L.z / Scalar(dim.z));
 
-        // decrease the number of bins if it exceeds the max
-        if (dim.x * dim.y * dim.z > m_max_cells)
-            {
-            Scalar scale_factor = powf(Scalar(m_max_cells) / Scalar(dim.x*dim.y*dim.z), Scalar(1.0)/Scalar(2.0));
-            dim.x = int(Scalar(dim.x)*scale_factor);
-            dim.y = int(Scalar(dim.y)*scale_factor);
-            }
+        // add cells up to the next integer to cover the whole ghost width
+        if (!box.getPeriodic().x)
+            dim.x += static_cast<int>(ceil(m_ghost_width.x/cell_size.x));
+
+        if (!box.getPeriodic().y)
+            dim.y += static_cast<int>(ceil(m_ghost_width.y/cell_size.y));
+
+        if (m_sysdef->getNDimensions() == 3 && !box.getPeriodic().z)
+            dim.z += static_cast<int>(ceil(m_ghost_width.z/cell_size.z));
         }
-    else
-        {
-        dim.z = roundDown((unsigned int)((L.z) / (m_nominal_width)), m_multiple);
+#endif
 
-        // add ghost layer if necessary
-        if (! box.getPeriodic().z)
-            dim.z += 2;
-
-        // decrease the number of bins if it exceeds the max
-        if (dim.x * dim.y * dim.z > m_max_cells)
-            {
-            Scalar scale_factor = powf(Scalar(m_max_cells) / Scalar(dim.x*dim.y*dim.z), Scalar(1.0)/Scalar(3.0));
-            dim.x = int(Scalar(dim.x)*scale_factor);
-            dim.y = int(Scalar(dim.y)*scale_factor);
-            dim.z = int(Scalar(dim.z)*scale_factor);
-            }
-        }
 
     // In extremely small boxes, the calculated dimensions could go to zero, but need at least one cell in each dimension
     // for particles to be in a cell and to pass the checkCondition tests.
@@ -166,6 +152,7 @@ uint3 CellList::computeDimensions()
         dim.y = 1;
     if (dim.z == 0)
         dim.z = 1;
+
     return dim;
     }
 
@@ -288,26 +275,40 @@ void CellList::initializeWidth()
     if (m_prof)
         m_prof->push("init");
 
+    // get the local box
+    const BoxDim& box = m_pdata->getBox();
+
+    // size the ghost layer width
+    m_ghost_width = make_scalar3(0.0, 0.0, 0.0);
+#ifdef ENABLE_MPI
+    if (m_comm)
+        {
+        Scalar ghost_width = m_comm->getGhostLayerMaxWidth();
+        if (ghost_width > Scalar(0.0))
+            {
+            if (!box.getPeriodic().x)
+                m_ghost_width.x = ghost_width;
+
+            if (!box.getPeriodic().y)
+                m_ghost_width.y = ghost_width;
+
+            if (m_sysdef->getNDimensions() == 3 && !box.getPeriodic().z)
+                m_ghost_width.z = ghost_width;
+            }
+        }
+#endif
+
     // initialize dimensions and width
     m_dim = computeDimensions();
 
-    const BoxDim& box = m_pdata->getBox();
+    // stash the current actual cell width
+    const Scalar3 L = box.getNearestPlaneDistance();
+    m_actual_width = make_scalar3((L.x + Scalar(2.0)*m_ghost_width.x) / Scalar(m_dim.x),
+                                  (L.y + Scalar(2.0)*m_ghost_width.y) / Scalar(m_dim.y),
+                                  (L.z + Scalar(2.0)*m_ghost_width.z) / Scalar(m_dim.z));
 
-    // the number of ghost cells along every non-periodic direction is two (one on each side)
-    if (m_sysdef->getNDimensions() == 2)
-        m_num_ghost_cells = make_uint3(box.getPeriodic().x ? 0 : 2,
-                                       box.getPeriodic().y ? 0 : 2,
-                                       0);
-    else
-        m_num_ghost_cells = make_uint3(box.getPeriodic().x ? 0 : 2,
-                                       box.getPeriodic().y ? 0 : 2,
-                                       box.getPeriodic().z ? 0 : 2);
-
-    // compute ghost layer width
-    Scalar3 L = box.getNearestPlaneDistance();
-    m_ghost_width = make_scalar3(L.x/(Scalar)(m_dim.x-m_num_ghost_cells.x)*(Scalar)(m_num_ghost_cells.x/2),
-                             L.y/(Scalar)(m_dim.y-m_num_ghost_cells.y)*(Scalar)(m_num_ghost_cells.y/2),
-                             L.z/(Scalar)(m_dim.z-m_num_ghost_cells.z)*(Scalar)(m_num_ghost_cells.z/2));
+    // signal that the width has changed
+    m_width_change();
 
     if (m_prof)
         m_prof->pop();
@@ -704,7 +705,6 @@ void export_CellList()
         ("CellList", init< boost::shared_ptr<SystemDefinition> >())
         .def("setNominalWidth", &CellList::setNominalWidth)
         .def("setRadius", &CellList::setRadius)
-        .def("setMaxCells", &CellList::setMaxCells)
         .def("setComputeTDB", &CellList::setComputeTDB)
         .def("setFlagCharge", &CellList::setFlagCharge)
         .def("setFlagIndex", &CellList::setFlagIndex)