+++ libhoomd/data_structures/ParticleData.h
@@ -110,8 +110,6 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 // Forward declaration of Profiler
 class Profiler;
 
-class WallData;
-
 // Forward declaration of RigidData
 class RigidData;
 
@@ -127,7 +125,8 @@ struct pdata_flag
         isotropic_virial=0,        //!< Bit id in PDataFlags for the isotropic virial
         potential_energy,          //!< Bit id in PDataFlags for the potential energy
         pressure_tensor,           //!< Bit id in PDataFlags for the full virial
-        rotational_kinetic_energy  //!< Bit id in PDataFlags for the rotational kinetic energy
+        rotational_kinetic_energy,  //!< Bit id in PDataFlags for the rotational kinetic energy
+        external_field_virial       //!< Bit id in PDataFlags for the external virial contribution of volume change
         };
     };
 
@@ -311,6 +310,7 @@ struct pdata_element
      - pdata_flag::potential_energy - specify that the potential energy .w component stored in the net force array
        (getNetForce) is valid
      - pdata_flag::pressure_tensor - specify that the full virial tensor is valid
+     - pdata_flag::external_field_virial - specify that an external virial contribution is valid
 
     If these flags are not set, these arrays can still be read but their values may be incorrect.